Частично сделала фильтры статистики
